Restar camara PMI

+++ Users/Super/Camera/showCamera.php
@@ -18,6 +18,8 @@ if(isset($_GET['action']) == 'delete'){
     $c= new Connector();
     $conn=$c->getCon();
     $query = mysqli_query($conn, "SELECT * FROM camara WHERE ns_cam='$id_delete'");
+    $row=mysqli_fetch_array($query);
+    $id_pmi=$row["id_pmi"];
     $queryCom = mysqli_query($conn, "SELECT * FROM comentarios WHERE identificador='$id_delete' and tabla='camara'");
 
     if(mysqli_num_rows($query) == 0 ){
@@ -27,15 +29,37 @@ if(isset($_GET['action']) == 'delete'){
         $delete = mysqli_query($conn, "DELETE FROM camara WHERE ns_cam='$id_delete'");
 
         if($delete){
-            $deleteCom = mysqli_query($conn, "DELETE FROM comentarios WHERE identificador='$id_delete' and tabla='camara'");
-            if($deleteCom){
-                mysqli_commit($conn);
-                header("Location: showCamera.php?e=1");
+            $PMI = new Connector();
+            $PMI->select("pmi","id_pmi",$id_pmi);
+            $queryp=$PMI->getQuery();
+            $row=mysqli_fetch_array($queryp);
+            $camaras=$row['num_cam'];
+            $camaras--;
+            if ($queryp) {
+                $PMI->update("pmi","num_cam='$camaras'","id_pmi",$id_pmi);
+                $queryp=$PMI->getQuery();
+                if($queryp){
+                    $deleteCom = mysqli_query($conn, "DELETE FROM comentarios WHERE identificador='$id_delete' and tabla='camara'");
+                    if($deleteCom){
+                        mysqli_commit($conn);
+                        header("Location: showCamera.php?e=1");
+                    }
+                    else{
+                        mysqli_rollback($conn);
+                        header("Location: showCamera.php?e=0");
+                    }
+                }
+                else{
+                    mysqli_rollback($conn);
+                    header("Location: showCamera.php?e=0");
+                }
             }
             else{
                 mysqli_rollback($conn);
                 header("Location: showCamera.php?e=0");
             }
+
+
         }else{
             mysqli_rollback($conn);
             header("Location: showCamera.php?e=0");
